create ReportingTools
